added slack channel tests
